+++ pyanalyze/__init__.py
@@ -7,14 +7,17 @@ from . import name_check_visitor
 from . import analysis_lib
 from . import annotations
 from . import arg_spec
-from .arg_spec import assert_is_value, dump_value
+from .implementation import assert_is_value, dump_value
 from . import asynq_checker
 from . import config
 from . import error_code
 from . import find_unused
 from .find_unused import used
+from . import implementation
 from . import method_return_type
 from . import node_visitor
+from . import safe
+from . import signature
 from . import stacked_scopes
 from . import test_config
 from . import tests
