+++ pyanalyze/test_attributes.py
@@ -1,5 +1,5 @@
 # static analysis: ignore
-from .arg_spec import assert_is_value
+from .implementation import assert_is_value
 from .value import KnownValue, MultiValuedValue, TypedValue, UNRESOLVED_VALUE
 from .test_node_visitor import assert_passes
 from .test_name_check_visitor import TestNameCheckVisitorBase
